test(export): pin the composed KMZ heading + up-axis orientation invariant
kmz.rs::heading_matches_ifc_convention and collada.rs::converts_yup_to_zup_and_centers
each pin one half of the KMZ rotation pipeline in isolation. Nothing asserted what
the two halves produce together, which is the only thing a Google Earth user
actually sees (#2554's bug class). Add composed_orientation_reproduces_ifc_bearing:
it runs the real ifc_angle_to_kml_heading + to_zup code paths for three IFC-encoded
bearings (30 deg, 200 deg in a different quadrant, and the identity axis (1,0) that
the original bug got wrong) and checks the recovered bearing of the exported
model's local +X axis matches the bearing the IFC file encoded.

Split collada.rs's inline #[cfg(test)] module out to collada_tests.rs (the repo's
_tests.rs sibling convention, ratchet-exempt) to make room without growing the
production file; refreshed its module-size-allowlist budget down from 635 to 430.

Verified non-vacuous: mutating to_zup's axis swap to (y, x, z), or flipping the
sign in ifc_angle_to_kml_heading's angle-to-heading formula, each independently
turns the new test red; both revert to green.
